+++ src/idc/writer/__init__.py
@@ -1,6 +1,7 @@
 from ._required_format_writer import RequiredFormatWriter
 from ._image_and_annotation_writer import ImageAndAnnotationWriter
 from ._data import DataWriter
+from ._metadata import MetaDataWriter
 from ._multi import MultiWriter
 from ._pyfunc import PythonFunctionWriter
 from ._send_email import SendEmail